backend/requirements.txt
Django>=4.2,<4.3
djangorestframework
dj-database-url
python-dotenv
django-environ
psycopg2-binary
openai>=1.12.0
requests
gunicorn
pgvector
